Cast ints up to longs before multiplying them into a long, or you get an overflow in the intermediate int

+++ com-mmxlabs-lng-scheduler/com.mmxlabs.scheduler.optimiser/src/main/java/com/mmxlabs/scheduler/optimiser/fitness/components/CharterCostFitnessComponent.java
@@ -61,7 +61,7 @@ public class CharterCostFitnessComponent<T> extends
 					final int arrivalTime = annotatedSequence.getAnnotation(lastElement, 
 							SchedulerConstants.AI_visitInfo, IPortVisitEvent.class).getStartTime(); //TODO check this is arrival time at last port
 					final int delta = arrivalTime - loadTime;
-					hireCost = delta * vessel.getVesselClass().getHourlyCharterPrice();
+					hireCost = ((long)delta) * ((long)vessel.getVesselClass().getHourlyCharterPrice());
 				}
 				
 				break;
